# plan-ai/backend/src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"GenericResponse": {
				"properties": {
					"message": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"message",
					"status"
				],
				"type": "object"
			},
			"_36_Enums.Role": {
				"type": "string",
				"enum": [
					"ADMIN",
					"PREMIUM",
					"CLIENT",
					"PENDING"
				]
			},
			"Role": {
				"$ref": "#/components/schemas/_36_Enums.Role"
			},
...
			"UserResponse": {
				"properties": {
					"id": {
						"type": "string"
					},
					"firebaseUid": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"avatarUrl": {
						"type": "string",
						"nullable": true
					},
					"googleId": {
						"type": "string",
						"nullable": true
					},
					"appleId": {
						"type": "string",
						"nullable": true
					},
					"microsoftId": {
						"type": "string",
						"nullable": true
					},
					"isGoogleAccount": {
						"type": "boolean"
					},
					"isAppleAccount": {
						"type": "boolean"
					},
					"isMicrosoftAccount": {
						"type": "boolean"
					},
					"role": {
						"$ref": "#/components/schemas/Role"
					},
					"hasCompletedOnboarding": {
						"type": "boolean"
					},
					"hasVoiceProfile": {
						"type": "boolean"
					},
					"voiceProfileUrl": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"firebaseUid",
					"email",
					"name",
					"avatarUrl",
					"googleId",
					"appleId",
					"microsoftId",
					"isGoogleAccount",
					"isAppleAccount",
					"isMicrosoftAccount",
					"role",
					"hasCompletedOnboarding",
					"hasVoiceProfile",
					"voiceProfileUrl"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ApiResponse_UserResponse_": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UserResponse"
							}
						],
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"ApiResponse__code-string__": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"code": {
								"type": "string"
							}
						},
						"required": [
							"code"
						],
						"type": "object",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
			"ApiResponse__customToken-string__": {
				"properties": {
					"message": {
						"type": "string"
					},
					"data": {
						"properties": {
							"customToken": {
								"type": "string"
							}
						},
						"required": [
							"customToken"
						],
						"type": "object",
						"nullable": true
					},
					"status": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"data",
					"status"
				],
				"type": "object"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "backend",
		"version": "1.0.0",
		"license": {
			"name": "BUSL-1.1"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/session/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserResponse_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericResponse"
								}
							}
						}
					}
				},
				"description": "Login with Firebase token.\nVerifies the token and returns user information.\nCreates a new user if one doesn't exist with the given Firebase UID.",
				"tags": [
					"Session"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"token": {
										"type": "string"
									},
									"uuid": {
										"type": "string"
									}
								},
								"required": [
									"token",
									"uuid"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
		"/api/session/me": {
			"get": {
				"operationId": "GetCurrentUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserResponse_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericResponse"
								}
							}
						}
					}
				},
				"description": "Get current user information.\nRequires authentication.",
				"tags": [
					"Session"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/session/me/voice-profile": {
			"post": {
				"operationId": "SaveVoiceProfile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_UserResponse_"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericResponse"
								}
							}
						}
					},
					"500": {
						"description": "Internal Server Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GenericResponse"
								}
							}
						}
					}
				},
				"description": "Save Voice Profile.\nUpdates the user to indicate they have completed voice enrollment.",
				"tags": [
					"Session"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": false,
					"content": {
						"multipart/form-data": {
							"schema": {
								"type": "object",
								"properties": {
									"voiceFile": {
										"type": "string",
										"format": "binary"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/session/desktop-token": {
			"post": {
				"operationId": "GetDesktopToken",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__code-string__"
								}
							}
						}
					}
				},
				"description": "Generate a Firebase Custom Token for the authenticated user.\nCalled by the web app after the user logs in, to hand off auth to the Electron desktop recorder.",
				"tags": [
					"Session"
				],
				"security": [
					{
						"BearerAuth": []
					}
				],
				"parameters": []
			}
		},
		"/api/session/desktop-exchange": {
			"post": {
				"operationId": "ExchangeDesktopCode",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse__customToken-string__"
								}
							}
						}
					}
				},
				"description": "Exchange the Short-Lived Code for a Firebase Custom Token.\nCalled securely behind the scenes by the Electron desktop recorder.",
				"tags": [
					"Session"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"properties": {
									"code": {
										"type": "string"
									}
								},
								"required": [
									"code"
								],
								"type": "object"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}